+++ client/android/app/src/main/assets/public/assets/all-wallets-7265faee.js
@@ -1,6 +1,0 @@
-import{m as o}from"./core-ef58bb7e.js";import"./wagmi-c444efa6.js";import"./vendor-c36eaa8e.js";import"./utils-617af415.js";import"./ui-c369d1f3.js";import"./events-91c3d7fe.js";import"./index.es-449de70f.js";const e=o`<svg fill="none" viewBox="0 0 24 24">
-  <path
-    style="fill: var(--wui-color-accent-100);"
-    d="M10.2 6.6a3.6 3.6 0 1 1-7.2 0 3.6 3.6 0 0 1 7.2 0ZM21 6.6a3.6 3.6 0 1 1-7.2 0 3.6 3.6 0 0 1 7.2 0ZM10.2 17.4a3.6 3.6 0 1 1-7.2 0 3.6 3.6 0 0 1 7.2 0ZM21 17.4a3.6 3.6 0 1 1-7.2 0 3.6 3.6 0 0 1 7.2 0Z"
-  />
-</svg>`;export{e as allWalletsSvg};