+++ client/android/app/src/main/assets/public/assets/arrow-bottom-13f3c400.js
@@ -1,8 +1,0 @@
-import{m as o}from"./core-ef58bb7e.js";import"./wagmi-c444efa6.js";import"./vendor-c36eaa8e.js";import"./utils-617af415.js";import"./ui-c369d1f3.js";import"./events-91c3d7fe.js";import"./index.es-449de70f.js";const a=o`<svg fill="none" viewBox="0 0 14 15">
-  <path
-    fill="currentColor"
-    fill-rule="evenodd"
-    d="M7 1.99a1 1 0 0 1 1 1v7.58l2.46-2.46a1 1 0 0 1 1.41 1.42L7.7 13.69a1 1 0 0 1-1.41 0L2.12 9.53A1 1 0 0 1 3.54 8.1L6 10.57V3a1 1 0 0 1 1-1Z"
-    clip-rule="evenodd"
-  />
-</svg>`;export{a as arrowBottomSvg};