+++ client/android/app/src/main/assets/public/assets/arrow-left-2c4d6468.js
@@ -1,8 +1,0 @@
-import{m as o}from"./core-ef58bb7e.js";import"./wagmi-c444efa6.js";import"./vendor-c36eaa8e.js";import"./utils-617af415.js";import"./ui-c369d1f3.js";import"./events-91c3d7fe.js";import"./index.es-449de70f.js";const a=o`<svg fill="none" viewBox="0 0 14 15">
-  <path
-    fill="currentColor"
-    fill-rule="evenodd"
-    d="M13 7.99a1 1 0 0 1-1 1H4.4l2.46 2.46a1 1 0 1 1-1.41 1.41L1.29 8.7a1 1 0 0 1 0-1.41L5.46 3.1a1 1 0 0 1 1.41 1.42L4.41 6.99H12a1 1 0 0 1 1 1Z"
-    clip-rule="evenodd"
-  />
-</svg>`;export{a as arrowLeftSvg};