+++ client/android/app/src/main/assets/public/assets/chevron-bottom-1d7cd8f2.js
@@ -1,8 +1,0 @@
-import{m as o}from"./core-ef58bb7e.js";import"./wagmi-c444efa6.js";import"./vendor-c36eaa8e.js";import"./utils-617af415.js";import"./ui-c369d1f3.js";import"./events-91c3d7fe.js";import"./index.es-449de70f.js";const n=o`<svg fill="none" viewBox="0 0 16 16">
-  <path
-    fill="currentColor"
-    fill-rule="evenodd"
-    d="M1.46 4.96a1 1 0 0 1 1.41 0L8 10.09l5.13-5.13a1 1 0 1 1 1.41 1.41l-5.83 5.84a1 1 0 0 1-1.42 0L1.46 6.37a1 1 0 0 1 0-1.41Z"
-    clip-rule="evenodd"
-  />
-</svg>`;export{n as chevronBottomSvg};