+++ client/android/app/src/main/assets/public/assets/close-c5189e6e.js
@@ -1,8 +1,0 @@
-import{m as o}from"./core-ef58bb7e.js";import"./wagmi-c444efa6.js";import"./vendor-c36eaa8e.js";import"./utils-617af415.js";import"./ui-c369d1f3.js";import"./events-91c3d7fe.js";import"./index.es-449de70f.js";const a=o`<svg fill="none" viewBox="0 0 16 16">
-  <path
-    fill="currentColor"
-    fill-rule="evenodd"
-    d="M2.54 2.54a1 1 0 0 1 1.42 0L8 6.6l4.04-4.05a1 1 0 1 1 1.42 1.42L9.4 8l4.05 4.04a1 1 0 0 1-1.42 1.42L8 9.4l-4.04 4.05a1 1 0 0 1-1.42-1.42L6.6 8 2.54 3.96a1 1 0 0 1 0-1.42Z"
-    clip-rule="evenodd"
-  />
-</svg>`;export{a as closeSvg};